refactor: quizzes > quiz 파일명 변경
[#131]
